+++ src/types/rapid-config.d.ts
@@ -1,8 +1,0 @@
-import {RapidEndpoint} from "../interfaces/IRapidEndpoint";
-
-export type RapidConfig = {
-    name: string,
-    endpoints: RapidEndpoint[],
-    port?: number,
-    prefix?: string
-}