+++ changelog.md
@@ -1,7 +1,14 @@
+### 1.0.1
+- Added missing type declarations
+- Small enhancements for the readme & overview page
+- Added new config option overviewPage > enable: to enable/disable the overview page
+- Added new config option overviewPage > theme: to set the theme of the overview page ("LIGHT" or "DARK")
+- type declarations didn't build correctly
+
 ### 1.0.0
 - get rid of config.json and add config object to RapidServer constructor
 - massive refactoring
-- fixed wrong error message when wrogn data type is provided
+- fixed wrong error message when wrong data type is provided
 - added error handling for invalid property names
 - updated readme